pyproject.toml
[project]
name = "cpt-to-soiltype"
version = "0.1.0"
description = "CPT to soiltype prediction"
authors = [{ name = "tfh", email = "tom.frode.hansen@ngi.no" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11.1"
dependencies = [
    "hydra-core>=1.3.2",
    "imblearn>=0.0",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "mlflow>=3.3.2",
    "nbformat>=5.10.4",
    "optuna>=4.5.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pydantic>=2.11.7",
    "pyod>=2.0.5",
    "statsmodels>=0.14.2",
    "hdbscan>=0.8.38",
    "umap-learn>=0.5.6",
    "rich>=14.1.0",
    "scikit-learn>=1.7.2",
    "streamlit>=1.49.1",
    "xgboost>=3.0.5",
    "ydata-profiling>=4.16.1",
    "featurewiz>=0.3.9",
    "lazytransform>=0.5.4",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "isort>=6.0.1",
    "mypy>=1.18.1",
    "ruff>=0.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cpt_to_soiltype"]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"


[tool.isort]
profile = "black"
force_single_line = false

[tool.ruff]
target-version = "py312"
line-length = 88

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".github",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
# Use single quotes for non-docstrings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Respect magic trailing comma
skip-magic-trailing-comma = false
# Use line feeds for line endings
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"


[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
ignore_missing_imports = true
files = ["src", "scripts"]
